+++ azure_monitor/tests/test_base_exporter.py
@@ -79,8 +79,7 @@ class TestBaseExporter(unittest.TestCase):
             "4321abcd-5678-4efa-8abc-1234567890ab",
         )
         self.assertEqual(
-            base.options.proxies,
-            '{"https":"https://test-proxy.com"}',
+            base.options.proxies, '{"https":"https://test-proxy.com"}',
         )
         self.assertEqual(base.options.storage_maintenance_period, 2)
         self.assertEqual(base.options.storage_max_size, 3)