+++ azure_monitor/CHANGELOG.md
@@ -3,7 +3,9 @@
 ## Unreleased
 
 - Remove dependency metrics from auto-collection
-- ([#92](https://github.com/microsoft/opentelemetry-azure-monitor-python/pull/92))
+  ([#92](https://github.com/microsoft/opentelemetry-azure-monitor-python/pull/92))
+- Change default local storage directory
+  ([#100](https://github.com/microsoft/opentelemetry-azure-monitor-python/pull/100))
 
 ## 0.3b.1
 Released 2020-05-21