+++ azure_monitor/CHANGELOG.md
@@ -2,10 +2,12 @@
 
 ## Unreleased
 
-- Remove request failed per second metrics from auto-collection
-  ([#102](https://github.com/microsoft/opentelemetry-azure-monitor-python/pull/102))
 - Remove dependency metrics from auto-collection
   ([#92](https://github.com/microsoft/opentelemetry-azure-monitor-python/pull/92))
+- Change default local storage directory
+  ([#100](https://github.com/microsoft/opentelemetry-azure-monitor-python/pull/100))
+- Remove request failed per second metrics from auto-collection
+  ([#102](https://github.com/microsoft/opentelemetry-azure-monitor-python/pull/102))
 
 ## 0.3b.1
 Released 2020-05-21