+++ azure_monitor/CHANGELOG.md
@@ -2,12 +2,12 @@
 
 ## Unreleased
 
-- Implement proxies in exporter configuration
-- ([#101](https://github.com/microsoft/opentelemetry-azure-monitor-python/pull/101))
 - Remove dependency metrics from auto-collection
   ([#92](https://github.com/microsoft/opentelemetry-azure-monitor-python/pull/92))
 - Change default local storage directory
   ([#100](https://github.com/microsoft/opentelemetry-azure-monitor-python/pull/100))
+- Implement proxies in exporter configuration
+  ([#101](https://github.com/microsoft/opentelemetry-azure-monitor-python/pull/101))
 - Remove request failed per second metrics from auto-collection
   ([#102](https://github.com/microsoft/opentelemetry-azure-monitor-python/pull/102))